Move db.py
