+++ tuva_dqi/app.py
@@ -4,8 +4,8 @@ import dash
 import dash_bootstrap_components as dbc
 from dash import Dash, html
 
+from db import init_db
 from pages.components import get_footer_component, get_navbar_component
-from pages.db import init_db
 
 app = Dash(
     __name__,
